+++ changelog.md
@@ -1,3 +1,12 @@
 # Smart BPM Trigger Changelog
+## v1.1.0 (2025-04-11)
+- Added individual BPM/BPB guideline color and width customization
+- Added orange/yellow/green guideline color and width customization
+- Increased width limit from 2 to 5
+- Replaced the snap to guideline button with a mod settings button (Suggested by [IOMAN](user:8265048))
+- Added filters to the snap to guideline button (Suggested by [IOMAN](user:8265048))
+- Added an option to distribute snapped objects across guidelines (Suggested by [Ishiguru](user:9757549))
+- Added a BPM spawn loop creator to spawn triggers in a rate of beats per minute (Suggested by [Brift](user:14114548))
+
 ## v1.0.0 (2025-02-18)
 - Initial release